removed older test file, replaced with newer version. tests repeat 10 times by default.
