+++ tests/test_material_files.py
@@ -1,15 +1,36 @@
+from random import choice, random, randrange, uniform
+import pytest
+
 from htsohm.material_files import random_number_density
 
+@pytest.fixture
+def LCs():
+    #==============================================================================
+    # default limits (see htsohm.sample.yaml)
+    LC_limits = [25.6, 51.2]
 
-def test_random_number_density():
-    """ Verify provided number_density_limits works by passing a suitably large pair of equal
-        limits and verifying that the random number is the only one possible. If the limits are
-        not enforced, there will be a range, and the test will fail.
-    """
-
-    lattice_constants = {
-        'a': 1,
-        'b': 1,
-        'c': 1,
+    #==============================================================================
+    # pick random lattice-constans within limits
+    LCs = {
+        'a'           : uniform(*LC_limits),
+        'b'           : uniform(*LC_limits),
+        'c'           : uniform(*LC_limits)
     }
-    assert random_number_density((100000,100000), lattice_constants) == 100000
+    LCs['volume'] = LCs['a'] * LCs['b'] * LCs['c']
+    return LCs
+
+@pytest.fixture
+def ND_limits():
+    #==========================================================================
+    # default limits (see htsohm.sample.yaml)
+    limits = [1.49e-05, 0.02122]
+    return limits
+
+def test_boundaries(LCs, ND_limits):
+#    for i in range(0, 10):
+     number_density = random_number_density(ND_limits, LCs) / LCs['volume']
+     assert number_density >= ND_limits[0] and number_density <= ND_limits[1]
+
+def test_two_site_constraint(LCs, ND_limits):
+#    for i in range(0, 10):
+    assert random_number_density(ND_limits, LCs) >= 2
